pyproject.toml
[project]
name = "hk-value-screener"
version = "0.1.0"
description = "A disciplined Hong Kong equities value screening system."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "akshare>=1.17.0",
  "pandas>=2.2.3",
  "pydantic>=2.11.0",
  "pypinyin>=0.55.0",
  "pyyaml>=6.0.2",
  "rich>=14.0.0",
  "typer>=0.16.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "ruff>=0.11.8",
]

[project.scripts]
hkvs = "hk_value_screener.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hk_value_screener"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
